pyproject.toml
[project]
name = "docs-agent"
version = "0.1.0"
description = "Contract editor with DeepSeek-assisted review and export"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.1",
    "openai>=1.55.0",
    "jinja2>=3.1.4",
    "python-docx>=1.1.2",
    "pypdf>=5.1.0",
    "pyyaml>=6.0.2",
    "beautifulsoup4>=4.12.3",
    "playwright>=1.49.0",
    "mammoth>=1.8.0",
]
